LayoutTests/fast/events/ios/key-command-italic-dispatches-keydown-expected.txt

Tests that pressing Command + I in a rich editing field dispatches a key down event. To run this test by hand, select all the text below and press Command + I.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.event.key is "i"
PASS window.event.metaKey is true
PASS successfullyParsed is true

TEST COMPLETE
